% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_project.R
\name{close_project}
\alias{close_project}
\title{Title}
\usage{
close_project(
  project_id,
  actual_end_ts = Sys.time(),
  userApiKey = getOption("userApiKey")
)
}
\arguments{
\item{project_id}{}

\item{actual_end_ts}{}

\item{userApiKey}{}
}
\value{

}
\description{
Title
}
