% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{extract_product_from_an_invoice}
\alias{extract_product_from_an_invoice}
\title{Title}
\usage{
extract_product_from_an_invoice(invoice)
}
\arguments{
\item{invoice}{an invoice}
}
\description{
Title
}
