% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2.R
\name{get_all}
\alias{get_all}
\title{Title}
\usage{
get_all(
  what,
  url = glue::glue("https://axonaut.com/api/v2/{what}"),
  userApiKey = getOption("userApiKey")
)
}
\arguments{
\item{what}{}

\item{url}{}

\item{userApiKey}{}
}
\value{

}
\description{
Title
}
