% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expenses.R
\name{get_arrayExpenses}
\alias{get_arrayExpenses}
\title{Title}
\usage{
get_arrayExpenses(accountApiKey = getOption("accountApiKey"),
  startDateTS = 0, endDateTS = 2451915527)
}
\arguments{
\item{accountApiKey}{token api}

\item{startDateTS}{start as epoch}

\item{endDateTS}{end as epoch}
}
\description{
Title
}
