% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{get_arrayInvoices}
\alias{get_arrayInvoices}
\title{Title}
\usage{
get_arrayInvoices(
  accountApiKey = getOption("accountApiKey"),
  startDateTS = 0,
  endDateTS = 2451915527
)
}
\arguments{
\item{accountApiKey}{token api}

\item{startDateTS}{start as epoch}

\item{endDateTS}{end as epoch}
}
\description{
Title
}
