% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project.R
\name{get_arrayProjects}
\alias{get_arrayProjects}
\title{Title}
\usage{
get_arrayProjects(accountApiKey = getOption("accountApiKey"))
}
\arguments{
\item{accountApiKey}{}
}
\value{

}
\description{
Title
}
