% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quotations.R
\name{get_arrayquotations}
\alias{get_arrayquotations}
\title{Title}
\usage{
get_arrayquotations(accountApiKey = getOption("accountApiKey"))
}
\arguments{
\item{accountApiKey}{token api}
}
\description{
Title
}
