% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{get_facture_detail}
\alias{get_facture_detail}
\title{Title}
\usage{
get_facture_detail(invoice = axonaut:::get_all_invoices())
}
\arguments{
\item{arrayInvoices}{arrayInvoices}
}
\description{
Title
}
