% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2.R
\name{create_companie}
\alias{create_companie}
\title{Title}
\usage{
create_companie(
  name,
  all_companies = get_all_companies(userApiKey = userApiKey),
  userApiKey = getOption("userApiKey")
)
}
\arguments{
\item{name}{}

\item{all_companies}{}

\item{userApiKey}{}
}
\value{

}
\description{
Title
}
