% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api2.R
\name{create_project}
\alias{create_project}
\title{Title}
\usage{
create_project(
  number,
  company_name,
  name = number,
  commentaire = name,
  company_id = get_company_id(company_name, userApiKey = userApiKey),
  all_project = get_all_project(userApiKey = userApiKey),
  userApiKey = getOption("userApiKey")
)
}
\arguments{
\item{number}{}

\item{company_name}{}

\item{name}{}

\item{commentaire}{}

\item{company_id}{}

\item{userApiKey}{}
}
\value{

}
\description{
Title
}
